+++ src/components/Header/Header.jsx
@@ -30,11 +30,11 @@ export default function Header({ onClick }) {
                 <li><Link to="/trackers">{isEn ? "Trackers" : "Трекери"}</Link></li>
                 <li><Link to="/analytics">{isEn ? "Analytics" : "Аналітика"}</Link></li>
                 <li><Link to="/projects">{isEn ? "Projects" : "Проєкти"}</Link></li>
-                <li><Link to="/tierlist">{isEn ? "TierList" : "Тірліст"}</Link></li>
+                {/* <li><Link to="/tierlist">{isEn ? "TierList" : "Тірліст"}</Link></li> */}
                 <li><Link to="/goalslist">{isEn ? "Goals" : "Цілі"}</Link></li>
                 <li><Link to="/calendar">{isEn ? "Calendar" : "Календар"}</Link></li>
+                <li><Link to="/help">{isEn ? "Instructions" : "Інструкція"}</Link></li>
                 <li><Link to="/settings">{isEn ? "Settings" : "Налаштування"}</Link></li>
-                <li><Link to="/help">{isEn ? "Допомога" : "Допомога"}</Link></li>
             </ul>
             <div className='burger-menu' onClick={onClick}>
                 <div className='line'></div>