+++ src/pages/Projects/components/ProjectStart.jsx
@@ -22,7 +22,7 @@ export default function TrackerStart({update}) {
         }
     }
     return (
-        <div className="trackers__add project__add">
+        <div className="trackers__add project__add newblock">
             <div className="trackers__info">
                 <input type="text" placeholder={isEn ? "Project Name" : "Назва проєкту"} value ={name} onChange={(e) => setName(e.target.value)}/>
             </div>